Return actual channel properties rather than callers in EnsureChannel/CreateChannel.
I've always thought it strange to pass back to the caller the properties
that the caller passed in.  I found that Mission Control on Maemo 5
actually expects the full set of properties to be returned or else it
will return whatever its gotten from multiple outstanding queries.  If
those queries don't return in time, the Call UI will report an error to
Mission Control which will close the channel and prevent the call from
completing.  To the user on Maemo 5 this is reported as "General Error".

For more information, see
https://bugs.maemo.org/show_bug.cgi?id=8980